requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# Authentication & Authorization
djangorestframework-simplejwt==5.3.0

# File handling and exports
openpyxl==3.1.2
python-pptx==0.6.22
WeasyPrint==60.2
Pillow==9.5.0

# Background tasks
django-q==1.3.9
redis==3.5.3

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Email
django-anymail==10.2

# File storage
django-storages==1.14.2
boto3==1.34.0

# Security
django-ratelimit==4.1.0
cryptography==43.0.3

# Monitoring and logging
sentry-sdk==1.38.0

# Data validation
cerberus==1.3.5

# Production server
gunicorn==21.2.0
dj-database-url==2.1.0
django-redis==5.4.0
